requirements.txt
dj-database-url==0.5.0
dj-email-url==0.1.0
Django==2.1
django-appconf==1.0.2
django-babel==0.6.2
django-cache-url==3.0.0
django-cors-headers==2.4.0
django-filter==2.0.0
django-heroku==0.3.1
django-jet==1.0.8
django-ranged-response==0.2.0
djangorestframework==3.9.0
djangorestframework-csv==2.1.0
google-api-python-client==1.4.1
oauth2client==2.2.0
request==1.0.2
requests==2.19.1
virtualenv==16.0.0
gunicorn==19.6.0

